se capstone/requirements.txt
flask
mysql-connector-python
qrcode[pil]
Pillow
scikit-learn
pandas
numpy
flask-socketio
eventlet
